+++ client/public/env.js
@@ -1,3 +1,3 @@
 window.env = {
-  "REACT_APP_MERCADOPAGO_KEY": "TEST-b1fe3820-a86d-4590-b6d0-f700d77ea02c"
+  "REACT_APP_MERCADOPAGO_KEY": "TEST-d6c63df8-4e5b-47af-9f16-0f4fd259b414"
 };
